+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-[Full Changelog](https://github.com/enderneko/Cell/compare/r215-release...6f2a3be83c89ed7cbba7ff24510677f99aca0855)
+[Full Changelog](https://github.com/enderneko/Cell/compare/r216-release...e602f935173f6534f03bbb951f113989e218da1b)
 
-- Fix Quick Assist
-- Fix an icon preview issue
+- Fix duration text visibility
+- Update zhTW
+- Update snippets